pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "dockerdo"
authors = [
    {name = "Stig-Arne Grönroos", email="stig.gronroos@gmail.com"}
]
readme = "README.rst"
license = "MIT"
dynamic = ["version", "description"]
classifiers = [ "Development Status :: 4 - Beta",
                "Intended Audience :: Science/Research",
                "Operating System :: OS Independent",
                "Programming Language :: Python",
                "Topic :: Scientific/Engineering",
                "Private :: Do Not Upload", ]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
]

[project.urls]
Documentation = "https://dockerdo.readthedocs.io"
Source = "https://github.com/waino/dockerdo"

[project.scripts]
dockerdo = "dockerdo:cli"

[tool.pytest.ini_options]
addopts = "-v --cov=dockerdo --cov-report html --ignore-glob docs/**"

[tool.ruff]
line-length = 120

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.6.12",
    "flit>=3.11.0",
    "ipython>=8.34.0",
    "mypy>=1.15.0",
    "mypy-extensions>=1.0.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.10",
    "sphinx>=8.1.3",
]
